+++ tests/TestCase/View/Helper/CookieHelperTest.php
@@ -37,16 +37,16 @@ class CookieHelperTest extends \PHPUnit\Framework\TestCase
     public function testExists()
     {
         $this->assertFalse($this->Cookie->exists('password'));
-        $this->Cookie->request()->cookies('foo','bar'); // Set in request
+        $this->Cookie->request()->cookies('foo', 'bar'); // Set in request
         $this->AssertTrue($this->Cookie->exists('foo'));
     }
     public function testRead()
     {
         // This is failing to decrypt when running all tests together
-        $_COOKIE['ez'] = 'T3JpZ2lu==.J4pNTvNegB4yYV1wgUK6eazwy+DSQhgAeOqmIIo7oeQ=';
+        $_COOKIE['ez'] = 'T3JpZ2lu==.Hh+COplGWNym2Xa+mhTS5hAh32OB4x6Rv0LflOdNx+cGLBJG9r0Ga1iP0XRPncRDPZ9jf5cv5IWoQm/JDNEIHg==';
         $_COOKIE['normal'] = 'hello';
         $this->Cookie = new CookieHelper(new View(new Controller(new Request(), new Response())));
-        $this->assertEquals('hello',$this->Cookie->read('normal'));
+        $this->assertEquals('hello', $this->Cookie->read('normal'));
         $this->assertEquals('bar', $this->Cookie->read('ez'));
     }
     public function testDelete()
@@ -60,7 +60,7 @@ class CookieHelperTest extends \PHPUnit\Framework\TestCase
     public function testdestroy()
     {
         $this->assertFalse($this->Cookie->exists('secret'));
-        $this->Cookie->request()->cookies('secret','bar'); // Set in request
+        $this->Cookie->request()->cookies('secret', 'bar'); // Set in request
         $this->assertTrue($this->Cookie->exists('secret'));
         $this->Cookie->destroy();
         $this->assertEquals([], $_COOKIE);